pyproject.toml
[project]
name = "curiopilot"
version = "0.1.0"
description = "Autonomous knowledge discovery agent (CurioPilot)"
readme = "README.md"
authors = [
    { name = "Egemen Çankaya", email = "egemencankaya14@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.12.3",
    "rich>=13.7.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "pyyaml>=6.0.1",
    "httpx>=0.27.0",
    "aiosqlite>=0.20.0",
    "tenacity>=8.2.3",
    "playwright>=1.40.0",
    "chromadb>=1.5.2",
    "networkx>=3.6.1",
    "langgraph>=1.0.10",
    "structlog>=24.1.0",
]

[project.scripts]
curiopilot = "curiopilot.cli:app"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.21.1",
]
